# xboost/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Xboost API
  description: |
    Xboost REST API仕様書
    
    X（旧Twitter）運用に必要な全機能を提供するSaaSプラットフォームのAPIです。
    
    ## 認証
    
    API認証は以下の2種類があります：
    
    ### 1. セッションベース認証（Webアプリ用）
    NextAuthを使用したセッションベース認証です。
    Cookieにセッション情報が含まれます。
    
    ### 2. APIキー認証（外部連携用）
    `X-API-Key` ヘッダーにAPIキーを設定してください。
    
    ```
    X-API-Key: your_api_key_here
    ```
    
    ## レート制限
    
    APIには以下のレート制限が適用されます：
    
    | プラン | 制限 |
    |--------|------|
    | Free | 100リクエスト/分 |
    | Pro | 1,000リクエスト/分 |
    | Enterprise | 10,000リクエスト/分 |
    
    レスポンスヘッダーで残りのリクエスト数を確認できます：
    - `X-RateLimit-Limit`: 制限値
    - `X-RateLimit-Remaining`: 残りリクエスト数
    - `X-RateLimit-Reset`: リセット時刻
    
    ## エラーレスポンス
    
    全てのエラーは以下の形式で返却されます：
    
    ```json
    {
      "error": "エラーメッセージ",
      "code": "ERROR_CODE",
      "details": {} // オプションの詳細情報
    }
    ```
    
  version: 1.0.0
  contact:
    name: Xboost Support
    email: support@xboost.now
    url: https://www.xboost.now/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://www.xboost.now/api
    description: 本番環境
  - url: http://localhost:3000/api
    description: 開発環境
paths:
...
  /payments/checkout:
    post:
      summary: Stripe Checkout Session作成
      description: サブスクリプションのCheckout Sessionを作成します
      tags:
        - Payments
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                priceId:
                  type: string
                trial:
                  type: boolean
                  default: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  url:
                    type: string
  /payments/portal:
    post:
      summary: Customer Portal作成
      description: Stripe Customer PortalのURLを生成します
      tags:
        - Payments
      security:
        - sessionAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
...
  /webhooks/stripe:
    post:
      summary: Stripe Webhook
      description: StripeからのWebhookイベントを処理します
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 成功
        '400':
          description: 署名検証失敗

  # AI
...
  /v1/analytics/ai-patterns:
    get:
      summary: AIパターン分析
      description: AIによる勝ちパターン分析結果を取得します
      tags:
        - Analytics
      security:
        - sessionAuth: []
      responses:
        '200':
          description: 成功
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.jsのセッションクッキー
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: APIキー認証
...